+++ src/Material/Material.cpp
@@ -1,6 +1,6 @@
 #include "Material.h"
 
-Material::Material(float E, float nu)
+Material::Material(double E, double nu)
   : ElasticModulus(E), poissonRatio(nu)
 {
   elasticMaterial();
@@ -14,10 +14,10 @@ void Material::elasticMaterial()
 {
   // this function will create a vector for material matrix
   // D11 D22 D33 D12 D13 D23
-  materialMatrix = new float[6];
-  float E1 = ElasticModulus*(1-poissonRatio)/((1-poissonRatio*2)*(1+poissonRatio));
-  float E2 = poissonRatio*E1/(1-poissonRatio);
-  float G  = ElasticModulus/(2*(1+poissonRatio));
+  materialMatrix = new double[6];
+  double E1 = ElasticModulus*(1-poissonRatio)/((1-poissonRatio*2)*(1+poissonRatio));
+  double E2 = poissonRatio*E1/(1-poissonRatio);
+  double G  = ElasticModulus/(2*(1+poissonRatio));
   materialMatrix[0] = E1;
   materialMatrix[1] = E1;
   materialMatrix[2] =  G;
