+++ src/StiffnessMatrix/StiffnessMatrixFirstOrder/StiffnessMatrixFirstOrder.h
@@ -24,9 +24,9 @@ public:
   virtual ~StiffnessMatrixFirstOrder();
   virtual Sparse& GetStiffnessMatrix() = 0;
   int GetStiffnessMatrixSize() override;
-  CUDA_HOSTDEV void stiffnessMatrixCalculation(unsigned int, unsigned int, float*, unsigned int*, float*, float*, float*, unsigned int* ,float*, unsigned int*, unsigned int*, unsigned int*, unsigned int*);
-  CUDA_HOSTDEV void stiffnessMatrixCalculation2(unsigned int, unsigned int, float*, unsigned int*, float*, float*, float* ,float*);
-  CUDA_HOSTDEV void constantCreator(unsigned int numberElement, float* c, float* x, float* y, unsigned int* mesh);
+  CUDA_HOSTDEV void stiffnessMatrixCalculation(unsigned int, unsigned int, double*, unsigned int*, double*, double*, double*, unsigned int* ,double*, unsigned int*, unsigned int*, unsigned int*, unsigned int*);
+  CUDA_HOSTDEV void stiffnessMatrixCalculation2(unsigned int, unsigned int, double*, unsigned int*, double*, double*, double* ,double*);
+  CUDA_HOSTDEV void constantCreator(unsigned int numberElement, double* c, double* x, double* y, unsigned int* mesh);
 protected:
 private:
