+++ src/Timer/Timer.cpp
@@ -16,6 +16,6 @@ Timer::~Timer()
 {
   end = std::chrono::high_resolution_clock::now();
   duration = end - start;
-  float ms = duration.count() * 1000.0f;
+  double ms = duration.count() * 1000.0f;
   std::cout << message << ms << "ms " << std::endl;
 };
